Restructure and remove rounding
