+++ internal/ping/handler.go
@@ -1,0 +1,10 @@
+package ping
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func Handler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprint(w, "PONG\n")
+}
